backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
python-jose[cryptography]
bcrypt
python-multipart
groq
pypdf
python-dotenv
sentence-transformers
pydantic[email]
arxiv
pdfplumber
